add hasMany table with modal mode

+++ resources/views/form/hasmanytable.blade.php
@@ -1,7 +1,13 @@
 @include("admin::form._header")
 
         <div id="has-many-{{$id}}">
-            <table class="table table-with-fields has-many-{{$id}} vertical-align-{{$verticalAlign}}">
+            <table class="table table-with-fields has-many-{{$id}} vertical-align-{{$verticalAlign}}"
+                @if(!empty($options['sortableHasManyTable']))
+                    data-sort-column="{{ $options['sortColumn'] }}"
+                    data-sort-group-column="{{ $options['sortGroupColumn'] }}"
+                    data-sort-with="{{ $options['sortWith'] }}"
+                @endif
+            >
                 <thead>
                 <tr>
                     @if(!empty($options['sortable']))
@@ -24,7 +30,24 @@
                     <tr class="has-many-{{$id}}-form fields-group">
 
                         @if(!empty($options['sortable']))
-                           <td width="20"><span class="icon-arrows-alt-v btn btn-light handle"></span></td>
+                            <td width="20">
+                                @if(!empty($options['sortableHasManyTable']))
+                                    @if(in_array($options['sortWith'], ['drag', 'all'], true))
+                                        <span class="icon-arrows-alt-v btn btn-light handle"></span>
+                                    @endif
+
+                                    @if(in_array($options['sortWith'], ['buttons', 'all'], true))
+                                        <button type="button" class="btn btn-light btn-sm has-many-sort-up">
+                                            <i class="icon-arrow-up"></i>
+                                        </button>
+                                        <button type="button" class="btn btn-light btn-sm has-many-sort-down">
+                                            <i class="icon-arrow-down"></i>
+                                        </button>
+                                    @endif
+                                @else
+                                    <span class="icon-arrows-alt-v btn btn-light handle"></span>
+                                @endif
+                            </td>
                         @endif
 
                         <?php $hidden = ''; ?>
@@ -54,19 +77,38 @@
             </table>
 
             <template class="{{$id}}-tpl">
-                <tr class="has-many-{{$id}}-form fields-group">
+                    <tr class="has-many-{{$id}}-form fields-group">
 
                     @if(!empty($options['sortable']))
-                        <td width="20"><span class="icon-arrows-alt-v btn btn-light handle"></span></td>
+                        <td width="20">
+                            @if(!empty($options['sortableHasManyTable']))
+                                @if(in_array($options['sortWith'], ['drag', 'all'], true))
+                                    <span class="icon-arrows-alt-v btn btn-light handle"></span>
+                                @endif
+
+                                @if(in_array($options['sortWith'], ['buttons', 'all'], true))
+                                    <button type="button" class="btn btn-light btn-sm has-many-sort-up">
+                                        <i class="icon-arrow-up"></i>
+                                    </button>
+                                    <button type="button" class="btn btn-light btn-sm has-many-sort-down">
+                                        <i class="icon-arrow-down"></i>
+                                    </button>
+                                @endif
+                            @else
+                                <span class="icon-arrows-alt-v btn btn-light handle"></span>
+                            @endif
+                        </td>
                     @endif
 
                     {!! $template !!}
 
-                    <td class="form-group">
-                        <div>
-                            <div class="remove btn btn-danger btn-sm pull-right">@include('admin::form._add_delete_button', ['type' => 'delete', 'defaultText' => trans('admin.remove'), 'defaultIcon' => 'icon-trash'])</div>
-                        </div>
-                    </td>
+                    @if($options['allowDelete'])
+                        <td class="form-group">
+                            <div>
+                                <div class="remove btn btn-danger btn-sm pull-right">@include('admin::form._add_delete_button', ['type' => 'delete', 'defaultText' => trans('admin.remove'), 'defaultIcon' => 'icon-trash'])</div>
+                            </div>
+                        </td>
+                    @endif
                 </tr>
             </template>
